feat: add db to userService.js, need to do the same in all services and update src in consequence
