+++ services/user/update_user_info-service/src/friendship.js
@@ -1,16 +1,6 @@
 import userService from "./userService.js";
 import { statusCode, returnMessages } from "./returnValues.js";
-
-const handleErrors = (fn) => async (req, res) => {
-	try {
-		await fn(req, res);
-	} catch (error) {
-		console.error(`Error in ${req.method} ${req.url}:`, error);
-		const status = error.status || statusCode.INTERNAL_SERVER_ERROR;
-		const message = error.message || returnMessages.INTERNAL_SERVER_ERROR;
-		res.code(status).send({ message });
-	}
-};
+import handleErrors from "./handleErrors.js";
 
 async function checkFriendshipStatus(user, friend) {
